Added some of the newer files to the ios-shared project. Fixed some copyrights, fixed CheckBox float.

+++ Source/UI/SDCheckbox.h
@@ -1,9 +1,10 @@
 //
 //  SDCheckbox.h
-//  walmart
+//  ios-shared
+
 //
 //  Created by Brandon Sneed on 1/15/14.
-//  Copyright (c) 2014 Walmart. All rights reserved.
+//  Copyright 2014 SetDirection. All rights reserved.
 //
 
 #import <UIKit/UIKit.h>
@@ -11,7 +12,7 @@
 @interface SDCheckbox : UIControl
 
 @property (nonatomic, strong) UIColor *checkboxColor UI_APPEARANCE_SELECTOR;
-@property (nonatomic, assign) float checkboxWidth UI_APPEARANCE_SELECTOR;
+@property (nonatomic, assign) CGFloat checkboxWidth UI_APPEARANCE_SELECTOR;
 @property (nonatomic, readonly) UILabel *textLabel;
 
 @property (nonatomic, assign) BOOL checked;
